Making the <nav> sticky
At first it wasn't working because the h1, h2 and <nav> were added within the <header> element. And the "stickiness" of <nav> was lost as soon as the <header> element was scrolled.
By removing the <header> and just having the <nav> with the <body> as the parent container it worked fine.